requirements.txt
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
scipy>=1.10.0
plotly>=5.13.0
python-binance>=1.0.17
fpdf>=1.7.2
pytest>=7.3.1